All occurances of multiplying the width and heights have been reverted to defaults and replaced with multiplying them by a varaible ("ITEM_SIZE_MULT") for better ease of mantainability or later changes.
A table ("SIZE_MULT_TOGGLE_TABLE") has been added with each category of items as keys with boolean values which controls which will have the size multiplier applies in their calculations. Just as a nice to have, not really seeing it as important functionality.

A few categories have had functions added to calculate their tetris sizes for consistency and code readability. This allows further logic to be added later if desired based on those items properties like the others.